requirements.txt
streamlit
pandas
numpy
matplotlib
pyarrow
google-api-python-client
google-auth
Pillow
